requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
aioredis==1.3.1
alembic==1.6.5
amqp==5.0.6
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
autoflake==1.4
Babel==2.9.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
flake8==3.9.2
greenlet==1.1.0
hiredis==2.0.0
idna==2.10
invoke==1.5.0
isort==5.9.1
kombu==5.1.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.8.1
prompt-toolkit==3.0.19
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
roboparse==0.0.2
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.19
sqlalchemy-mixins==1.4
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
